Prevent fatal error when it gets a 403 Error
This commit introduces a handler for `client.getWorkers` when it receives an error argument.

This way it won't break JS from running, but instead will just report the error and continue.

This is not the best way to fix this problem, but this will prevent false statements on CI runs, creating warnings rather than fatal errros.