Fix crash from property 'browser_version' undefined
When no match is found, the process would just crash. Instead, recognise
this as a normal end-user scenario and point them to the list of
available workers.

In my experience the most common way to encounter this issue is when
additional decimal places are required by the API, e.g. "safari"
"9" results in a crash, because the available workers are described
as "9.1" instead. Given that most browers are described without a
decimal place, this is an easy mistake to make.

Another way to get this is when there is a misspelled browser name.

Before:

```
TypeError: Cannot read property 'browser_version' of undefined
    at browserstack-runner/bin/cli.js
```

After:

```
Error: No desktop match found for {"browser":"safari","browser_version":"9"}
Check https://www.browserstack.com/list-of-browsers-and-platforms/js_testing
    at populateOsAndOsVersion (browserstack-runner/lib/configParser.js)

```

Fixes https://github.com/browserstack/browserstack-runner/issues/243.